feat: use post word hook
